Adjust EOL to be one year from the last release
Two years was _way_ too permissive.  Two years from the "initial" release would be fine, but from every subsequent is just too much.

Most maintenance releases are one year or less apart, barring the single outlier of 4.7.3 to 4.7.4:

	gcc-4.7.0/	22-Mar-2012 05:36
	gcc-4.7.1/	14-Jun-2012 08:36
	gcc-4.7.2/	20-Sep-2012 10:58
	gcc-4.7.3/	11-Apr-2013 05:41
	gcc-4.7.4/	12-Jun-2014 14:36

	gcc-4.8.0/	22-Mar-2013 15:18
	gcc-4.8.1/	31-May-2013 06:40
	gcc-4.8.2/	16-Oct-2013 05:06
	gcc-4.8.3/	22-May-2014 07:55
	gcc-4.8.4/	19-Dec-2014 08:06
	gcc-4.8.5/	23-Jun-2015 11:36

	gcc-4.9.0/	22-Apr-2014 07:16
	gcc-4.9.1/	16-Jul-2014 09:16
	gcc-4.9.2/	30-Oct-2014 06:01
	gcc-4.9.3/	26-Jun-2015 16:30

	gcc-5.1.0/	22-Apr-2015 06:41

	gcc-5.2.0/	16-Jul-2015 08:21

(This does also mean that 4.7 is EOL, so if this merges, I'll be removing it from the official build rotation promptly.)