Add 11.1
This uses Debian Bullseye to work around compilation errors (and because it's going to be around a while so we might as well start it on "Debian Next" which is due to be released Soon™).

This also adds explicit "Debian Suite" aliases to the tagging so that it's more clear on https://hub.docker.com/_/gcc which base is in use.