Switch to upstream mirror for scraping version numbers
Currently, we have a 12.4 release that's published, but not (yet) on the kernel.org mirror, so our version scraping fails to pick it up.

While I feel less bad about scraping the kernel.org mirror load-wise, I think it's probably reasonable for our only-a-couple-times-daily version scraping script to poke the official upstream server to find the official upstream releases without relying on them having propagated to the mirrors yet (especially since it's not downloading the large bits -- just the HTML file listing).

(Also add it as a very low priority fallback during download just in case it's the literal only site which has the version we're updating to.)